Use s instead of zs for recon loss in training set
In case of binary missing subgroup, this is equivalent to setting
zs_dim=0, because zs will always be a constant. However, in contrast to
conditionally setting zs_dim to 0, the approach in this PR also scales
to non-binary cases, and might also benefit the subgroup bias setting.

The s values are one-hot encoded if they are non-binary. This requires
`zs_dim == s_dim`, for which I added a check.